Fixed extra packages

+++ wbdv-s1-2020-jh-nd-server-java/src/main/java/com/example/myapp/services/UserService.java
@@ -13,7 +13,7 @@ public class UserService {
     return repository.save(user);
   }
 
-  public User findUserById(Integer id) {
+  public User findUserById(Long id) {
     return repository.findUserById(id);
   }
